Moved some pylearn2 specific training code into that backend.
